Update and link scenes

+++ Assets/Scripts/Bullet.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Bullet : MonoBehaviour
 {
-    public float bulletDamage;
+    public int bulletDamage;
     private void OnCollisionEnter2D(Collision2D collision)
     {
         Destroy(gameObject);
