+++ Assets/Scripts/Shooting.cs
@@ -50,7 +50,7 @@ public class Shooting : MonoBehaviour
                 currentAttackTimer = 0f;
 
                 GameObject bullet = Instantiate(bulletPrefab, firePoint.position, firePoint.rotation);
-                bullet.GetComponent<Bullet>().bulletDamage = float.Parse(attackValue.text);
+                bullet.GetComponent<Bullet>().bulletDamage = int.Parse(attackValue.text);
                 Rigidbody2D rbBullet = bullet.GetComponent<Rigidbody2D>();
                 rbBullet.AddForce(firePoint.up * bulletForce, ForceMode2D.Impulse);
             }
